+++ gltf-json/src/animation.rs
@@ -168,12 +168,12 @@ pub struct Sampler {
 }
 
 impl Validate for Animation {
-    fn validate_minimally<P, R>(&self, root: &Root, path: P, report: &mut R)
+    fn validate<P, R>(&self, root: &Root, path: P, report: &mut R)
     where
         P: Fn() -> Path,
         R: FnMut(&Fn() -> Path, Error),
     {
-        self.samplers.validate_minimally(root, || path().field("samplers"), report);
+        self.samplers.validate(root, || path().field("samplers"), report);
         for (index, channel) in self.channels.iter().enumerate() {
             if channel.sampler.value() as usize >= self.samplers.len() {
                 let path = || path().field("channels").index(index).field("sampler");