+++ gltf-json/src/extras.rs
@@ -38,13 +38,7 @@ impl fmt::Display for Void {
 
 #[cfg(feature = "extras")]
 impl Validate for ::std::boxed::Box<RawValue> {
-    fn validate_minimally<P, R>(&self, _: &Root, _: P, _: &mut R)
-        where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
-    {
-        // nop
-    }
-
-    fn validate_completely<P, R>(&self, _: &Root, _: P, _: &mut R)
+    fn validate<P, R>(&self, _: &Root, _: P, _: &mut R)
         where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
     {
         // nop