+++ gltf-json/src/image.rs
@@ -48,12 +48,4 @@ pub struct Image {
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct MimeType(pub String);
 
-impl Validate for MimeType {
-    fn validate_completely<P, R>(&self, _: &Root, path: P, report: &mut R)
-        where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
-    {
-        if !VALID_MIME_TYPES.contains(&self.0.as_str()) {
-            report(&path, Error::Invalid);
-        }
-    }
-}
+impl Validate for MimeType {}