+++ gltf-json/src/validation.rs
@@ -4,23 +4,10 @@ use std::hash::Hash;
 
 use crate::{Path, Root};
 
-/// Trait for validating glTF JSON data against the 2.0 specification.
+/// Trait for validating glTF JSON data so that the library can function without panicking.
 pub trait Validate {
-    /// Validates only the invariants required for the library to function safely.
-    fn validate_minimally<P, R>(&self, _root: &Root, _path: P, _report: &mut R)
-    where
-        P: Fn() -> Path,
-        R: FnMut(&Fn() -> Path, Error),
-    {
-        // nop
-    }
-
-    /// Validates the data against the glTF 2.0 specification.
-    ///
-    /// # Notes
-    ///
-    /// The caller must also call `validate_minimally()` for full validation.
-    fn validate_completely<P, R>(&self, _root: &Root, _path: P, _report: &mut R)
+    /// Validates the invariants required for the library to function safely.
+    fn validate<P, R>(&self, _root: &Root, _path: P, _report: &mut R)
     where
         P: Fn() -> Path,
         R: FnMut(&Fn() -> Path, Error),
@@ -103,7 +90,7 @@ impl<T: Default> Default for Checked<T> {
 }
 
 impl<T> Validate for Checked<T> {
-    fn validate_minimally<P, R>(&self, _root: &Root, path: P, report: &mut R)
+    fn validate<P, R>(&self, _root: &Root, path: P, report: &mut R)
         where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
     {
         match *self {
@@ -114,57 +101,32 @@ impl<T> Validate for Checked<T> {
 }
 
 impl<K: Eq + Hash + ToString + Validate, V: Validate> Validate for HashMap<K, V> {
-    fn validate_minimally<P, R>(&self, root: &Root, path: P, report: &mut R)
+    fn validate<P, R>(&self, root: &Root, path: P, report: &mut R)
         where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
     {
         for (key, value) in self.iter() {
-            key.validate_minimally(root, || path().key(&key.to_string()), report);
-            value.validate_minimally(root, || path().key(&key.to_string()), report);
-        }
-    }
-
-    fn validate_completely<P, R>(&self, root: &Root, path: P, report: &mut R)
-        where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
-    {
-        for (key, value) in self.iter() {
-            key.validate_completely(root, || path().key(&key.to_string()), report);
-            value.validate_completely(root, || path().key(&key.to_string()), report);
+            key.validate(root, || path().key(&key.to_string()), report);
+            value.validate(root, || path().key(&key.to_string()), report);
         }
     }
 }
 
 impl<T: Validate> Validate for Option<T> {
-    fn validate_minimally<P, R>(&self, root: &Root, path: P, report: &mut R)
+    fn validate<P, R>(&self, root: &Root, path: P, report: &mut R)
         where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
     {
         if let Some(value) = self.as_ref() {
-            value.validate_minimally(root, path, report);
-        }
-    }
-
-    fn validate_completely<P, R>(&self, root: &Root, path: P, report: &mut R)
-        where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
-    {
-        if let Some(value) = self.as_ref() {
-            value.validate_completely(root, path, report);
+            value.validate(root, path, report);
         }
     }
 }
 
 impl<T: Validate> Validate for Vec<T> {
-    fn validate_minimally<P, R>(&self, root: &Root, path: P, report: &mut R)
+    fn validate<P, R>(&self, root: &Root, path: P, report: &mut R)
         where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
     {
         for (index, value) in self.iter().enumerate() {
-            value.validate_minimally(root, || path().index(index), report);
-        }
-    }
-
-    fn validate_completely<P, R>(&self, root: &Root, path: P, report: &mut R)
-        where P: Fn() -> Path, R: FnMut(&Fn() -> Path, Error)
-    {
-        for (index, value) in self.iter().enumerate() {
-            value.validate_completely(root, || path().index(index), report);
+            value.validate(root, || path().index(index), report);
         }
     }
 }